Specify int sizes in yokkaichi_structs.h

+++ yokkaichi_c/yokkaichi_structs.h
@@ -1,16 +1,17 @@
 #ifndef YOKKAICHI_STRUCTS_H
 #define YOKKAICHI_STRUCTS_H
 
+#include <stdint.h>
 #include "yokkaichi_constants.h"
 
 
 typedef struct {
     char ip[IP_LENGHT];
-    int port;
+    uint16_t port;
 } MinecraftServer; /* TODO: Rename this struct, a different struct will use this name in the future */
 
 typedef struct {
-    int threadId;
+    uint16_t threadId;
     MinecraftServer *server;
 } ThreadArgs;
